# cmd/drivarrd/web/openapi.yaml
openapi: 3.1.0
info:
  title: Drivarr API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post: {summary: Create a local session, responses: {'200': {description: Signed in}, '401': {description: Invalid credentials}}}
...
  /jobs/{id}/pause:
    post:
      summary: Pause a checkpointable test
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Pause requested}}
  /jobs/{id}/resume:
    post:
      summary: Resume a paused or interrupted test
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Resume requested}}
  /jobs/{id}/cancel:
    post:
      summary: Cancel a test
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Cancellation requested}}
  /jobs/{id}/report:
    post:
      summary: Generate an immutable PDF report
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'201': {description: Report created}}
  /devices/{id}/report:
    post:
      summary: Generate an immutable full-drive PDF containing every retained test
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'201': {description: Full drive report created}, '400': {description: No retained tests}}
...